Make toolbar work.
Turns out, the documentation I was following was broken and incorrectly said the toolbar was not supported natively (and required the Toolbar class instead).
But the v7 compatibility library already contains the right logic to make this work.

Since there are no actions associated with the toolbar, it's only item is disabled for now.